Hide admin recipe if not logged in
