+++ src/global/components/sign-in-form.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function signInForm(props) {
+    return <div>
+        Coming soon...
+    </div>
+}
